Patch printer driver to build with modern GCC (14.2.1)
- Fix C++ language standard compatibility issues:
  * Remove deprecated throw() dynamic exception specifications
  * Fix pointer-to-int casting using uintptr_t for 64-bit compatibility
  * Resolve string literal to char* conversion issues using const_cast
  * Fix ambiguous byte type conflicts with std::byte
  * Fix stream comparison loops (0 < stream.getline() -> stream.getline())
  * Fix narrowing conversion errors for hex literals (char -> unsigned char)
  * Fix void*/std::string* return type issues (return false -> return nullptr)
  * Remove deprecated register keyword

- Fix Autotools compatibility:
  * Update setupOmni script for newer autotools versions
  * Fix AM_INIT_AUTOMAKE duplicate expansion
  * Fix Makefile.am linking issues

- Fix library compatibility:
  * Create glib-config compatibility layer for glib-2.0
  * Fix XML library linking order
  * Fix gmodule-2.0 linking issues

- Fix source file copying issue in Epson High Res ESCP2 Blitter

The project now builds successfully with GCC 14.2.1 using ./setupOmni --enable-devicecompile

+++ DeviceTester5.cpp
@@ -20,6 +20,7 @@
 #include <Device.hpp>
 #include <Omni.hpp>
 #include <JobProperties.hpp>
+#include <cstdint>
 
 Device *
 loadDevice (PSZCRO    pszFullDeviceName,
@@ -62,7 +63,7 @@ loadDevice (PSZCRO    pszFullDeviceName,
                     "newDeviceW_JopProp_Advanced",
                     (gpointer *)&pfnNewDeviceWArgs);
 
-   if (DebugOutput::shouldOutputDeviceTester ()) DebugOutput::getErrorStream () << __FUNCTION__ << ": pfnNewDeviceWArgs = 0x" << std::hex << (int)pfnNewDeviceWArgs << std::dec << std::endl;
+   if (DebugOutput::shouldOutputDeviceTester ()) DebugOutput::getErrorStream ()       << __FUNCTION__ << ": pfnNewDeviceWArgs = 0x" << std::hex << (uintptr_t)pfnNewDeviceWArgs << std::dec << std::endl;
 
    if (!pfnNewDeviceWArgs)
       return 0;
